Added bottom NavigationBar; Added DashedInputText file; changed the border of the gender selector
