+++ petJournal/app/src/main/java/com/soujunior/petjournal/ui/components/InputText.kt
@@ -133,11 +133,6 @@ fun InputText(
                 ),
                 modifier = modifier
                     .height(40.dp)
-//                    .dashedBorder(
-//                        shape = RoundedCornerShape(10.dp),
-//                        isError = isError,
-//                        isSelected = false
-//                    )
                     .onFocusChanged {
                         inFocus = if (it.hasFocus)
                             it.hasFocus
@@ -165,148 +160,5 @@ fun InputText(
 
 
 
-
-@Composable
-fun DashedInputText(
-    modifier : Modifier = Modifier
-        .padding(5.dp)
-        .fillMaxWidth(),
-    textInputModifier: Modifier = Modifier,
-    placeholderText: String = "Placeholder",
-    titleText: String = "Title",
-    textValue: String,
-    isPassword: Boolean = false,
-    isError: Boolean = false,
-    textError: List<String>? = null,
-    onEvent: (String) -> Unit,
-    keyboardOptions: KeyboardOptions = KeyboardOptions.Default.copy(keyboardType = KeyboardType.Text),
-    visualTransformation: VisualTransformation = VisualTransformation.None){
-    var inFocus by remember { mutableStateOf(false) }
-    var showPassword by remember { mutableStateOf(false) }
-
-    Column(modifier = modifier) {
-        Row {
-            Text(
-                text = titleText,
-                textAlign = TextAlign.Start,
-                color = MaterialTheme.colorScheme.onBackground,
-                style = MaterialTheme.typography.bodyMedium,
-                fontSize = 15.sp,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(start = 10.dp, bottom = 5.dp)
-            )
-        }
-
-        Row{
-            BasicTextField(
-                modifier = textInputModifier
-                    .background(
-                        MaterialTheme.colorScheme.surface
-                    )
-                    .fillMaxWidth()
-                    .padding(5.dp)
-                    .height(40.dp)
-                    .drawBehind {
-                        val stroke = Stroke(
-                            width = 1.dp.toPx(),
-                            pathEffect = PathEffect.dashPathEffect(
-                                intervals = floatArrayOf(8.dp.toPx(), 8.dp.toPx(), 0f)
-                            )
-                        )
-                        drawRoundRect(
-                            color = if (isError) Color.Transparent else Color.Black,
-                            style = stroke,
-                            cornerRadius = CornerRadius(10.dp.toPx())
-                        )
-
-                }
-                .border(2.dp, if(isError) MaterialTheme.colorScheme.error else Color.Transparent, shape = RoundedCornerShape(10.dp))
-                .clip(RoundedCornerShape(10.dp)),
-                value = textValue,
-                onValueChange = {text -> onEvent(text)},
-                singleLine = true,
-                textStyle = TextStyle(
-                    fontSize = 15.sp,
-                    color = MaterialTheme.colorScheme.onSurface
-                ),
-                maxLines = 1,
-                visualTransformation =
-                if (isPassword) {
-                    if (showPassword) VisualTransformation.None
-                    else PasswordVisualTransformation()
-                } else visualTransformation,
-                keyboardOptions = keyboardOptions,
-                decorationBox = {innerTextField ->
-                    Row(
-                        modifier,
-                        verticalAlignment = Alignment.CenterVertically
-                    ){
-                        Box(
-                            modifier.weight(1f)){
-                            if(textValue.isEmpty()) //Placeholder
-                                Text(
-                                    modifier = modifier.padding(10.dp),
-                                    text = placeholderText,
-                                    style = MaterialTheme.typography.bodyMedium,
-                                    color = MaterialTheme.colorScheme.onSurface,
-                                    fontSize = 15.sp
-                                )
-                            else
-                                Text(
-                                    modifier = modifier.padding(10.dp),
-                                    text = textValue
-                                )
-                        }
-                            //Logica do Icone
-                            if (isPassword) {
-                                val iconResource =
-                                    if (showPassword) R.drawable.eye_visibility_on else R.drawable.eye_visibility_off
-                                val contentDescription =
-                                    if (showPassword) "Ocultar senha" else "Mostrar senha"
-
-                                IconButton(onClick = { showPassword = !showPassword }) {
-                                    Icon(
-                                        painter = painterResource(id = iconResource),
-                                        contentDescription = contentDescription,
-                                        tint = MaterialTheme.colorScheme.outline
-                                    )
-                                }
-                            } else if(isError) {
-                                val iconResource = R.drawable.icone_erro
-                                val contentDescription = "Erro"
-
-                                Icon(
-                                    painter = painterResource(id = iconResource),
-                                    contentDescription = contentDescription,
-                                    tint = Color.Unspecified,
-                                    modifier = Modifier.padding(10.dp)
-                                )
-                            }
-                            else {
-                                null
-                            }
-                        }
-                }
-            )
-        }
-        Row {
-            if(textError != null) {
-                textError.forEach {
-                    AlertText(textMessage = it, modifier = Modifier.padding(10.dp))
-                }
-            }
-            else{
-                Text(
-                    "*Campo Obrigatório.",
-                    color = MaterialTheme.colorScheme.outline,
-                    modifier = Modifier.padding(10.dp),
-                    fontSize = 15.sp
-                )
-            }
-        }
-    }
-
-}
