+++ petJournal/app/src/main/java/com/soujunior/petjournal/ui/components/RoundedSquare.kt
@@ -2,6 +2,7 @@ package com.soujunior.petjournal.ui.components
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
+import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -17,7 +18,12 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.drawBehind
+import androidx.compose.ui.geometry.CornerRadius
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.PathEffect
+import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.Dp
@@ -93,13 +99,22 @@ fun RoundedSquare(
                     bottomEnd = bottomRightRadius
                 )
             )
-            .dashedBorder(
-                shape = RoundedCornerShape(30.dp),
-                isError = false,
-                isSelected = isSelected,
-                strokeWidth = 1.5.dp,
-                selectedColor = selectedColor
-            )
+            .drawBehind {
+                val stroke = Stroke(
+                    width = 1.5.dp.toPx(),
+                    pathEffect = PathEffect.dashPathEffect(
+                        intervals = floatArrayOf(8.dp.toPx(), 8.dp.toPx(), 0f)
+                    )
+                )
+                drawRoundRect(
+                    color = if(isSelected) Color.Transparent else Color.Black,
+                    style = stroke,
+                    cornerRadius = CornerRadius(30.dp.toPx())
+                )
+
+            }
+            .border(2.dp, if(isSelected) MaterialTheme.colorScheme.primary else Color.Transparent, shape = RoundedCornerShape(30.dp))
+            .clip(RoundedCornerShape(30.dp))
     )
     {
         Column(
