+++ petJournal/app/src/main/java/com/soujunior/petjournal/ui/navigation/NavHostElements.kt
@@ -71,5 +71,6 @@ fun NavTestScreen(){
     NavHost(navController = navController, startDestination = "pets/nameAndGender") {
         composable("pets/nameAndGender"){
             PetNameAndGenderScreen(petSpecie = "Gato", navController = navController)}
+
     }
 }
